+++ examples/pyplot/plot_density3d.py
@@ -14,8 +14,8 @@ pts = Points(p, alpha=0.5)
 vol = pts.density() # density() returns a Volume
 vol.cmap('Dark2').alpha([0.1,1])
 
-r = precision(vol.metadata['radius'], 2) # retrieve automatic radius value
-vol.add_scalarbar3d(title='Density (counts in r_s ='+r+')', italic=1)
+r = precision(vol.metadata['radius'][0], 2) # retrieve automatic radius value
+vol.add_scalarbar3d(title=f'Density (counts in r_s ={r})', italic=1)
 
 show(pts, vol, __doc__, axes=True).close()